refactor: move union-code to union package
