+++ extras/src/main/scala/io/github/martinhh/derived/extras/union/TypedTypeClass.scala
@@ -1,4 +1,4 @@
-package io.github.martinhh.derived.extras
+package io.github.martinhh.derived.extras.union
 
 import scala.reflect.TypeTest
